+++ src-ui/app/app.module.ts
@@ -156,6 +156,9 @@ import { SleepModeEnableOnHeartRateCalmPeriodAutomationService } from './service
 import { HeartRateChartComponent } from './views/dashboard-view/views/sleep-detection-view/heart-rate-calm-period-enable-sleepmode-modal/heart-rate-chart/heart-rate-chart.component';
 import { StartWithSteamVRHowToModalComponent } from './views/dashboard-view/views/settings-view/settings-general-tab/confirm-modal/start-with-steamvr-how-to-modal.component';
 import { QuitWithSteamVRService } from './services/quit-with-steamvr.service';
+import { VRChatMicMuteAutomationService } from './services/osc-automations/vrchat-mic-mute-automation.service';
+import { MiscTestingComponent } from './components/developer-debug-modal/misc-testing/misc-testing.component';
+import { VRChatMicMuteAutomationsViewComponent } from './views/dashboard-view/views/vrchat-mic-mute-automations-view/vrchat-mic-mute-automations-view.component';
 
 [localeEN, localeFR, localeCN_TW, localeNL, localeKO, localeJP, localeES, localeID].forEach(
   (locale) => registerLocaleData(locale)
@@ -246,6 +249,8 @@ export function createTranslateLoader(http: HttpClient) {
     HeartRateCalmPeriodEnableSleepModeModalComponent,
     HeartRateChartComponent,
     StartWithSteamVRHowToModalComponent,
+    MiscTestingComponent,
+    VRChatMicMuteAutomationsViewComponent,
   ],
   imports: [
     CommonModule,
@@ -324,6 +329,7 @@ export class AppModule {
     // OSC automations
     private oscGeneralAutomationsService: OscGeneralAutomationsService,
     private sleepingAnimationsAutomationService: SleepingAnimationsAutomationService,
+    private vrchatMicMuteAutomationService: VRChatMicMuteAutomationService,
     // Status automations
     private statusChangeForPlayerCountAutomationService: StatusChangeForPlayerCountAutomationService,
     // Invite automations
@@ -412,6 +418,7 @@ export class AppModule {
           // OSC automations
           this.oscGeneralAutomationsService.init(),
           this.sleepingAnimationsAutomationService.init(),
+          this.vrchatMicMuteAutomationService.init(),
           // Status automations
           this.statusChangeForPlayerCountAutomationService.init(),
           // Invite automations