+++ src-ui/app/models/automations.ts
@@ -26,6 +26,7 @@ export type AutomationType =
   // OSC AUTOMATIONS
   | 'OSC_GENERAL'
   | 'SLEEPING_ANIMATIONS'
+  | 'VRCHAT_MIC_MUTE_AUTOMATIONS'
   // STATUS AUTOMATIONS
   | 'CHANGE_STATUS_BASED_ON_PLAYER_COUNT'
   // INVITE AUTOMATIONS
@@ -69,6 +70,7 @@ export interface AutomationConfigs {
   // OSC AUTOMATIONS
   OSC_GENERAL: OscGeneralAutomationConfig;
   SLEEPING_ANIMATIONS: SleepingAnimationsAutomationConfig;
+  VRCHAT_MIC_MUTE_AUTOMATIONS: VRChatMicMuteAutomationsConfig;
   // STATUS AUTOMATIONS
   CHANGE_STATUS_BASED_ON_PLAYER_COUNT: ChangeStatusBasedOnPlayerCountAutomationConfig;
   // INVITE AUTOMATIONS
@@ -216,6 +218,15 @@ export interface SleepingAnimationsAutomationConfig extends AutomationConfig {
   footLockReleaseWindow: number;
 }
 
+export type VRChatVoiceMode = 'TOGGLE' | 'PUSH_TO_MUTE';
+
+export interface VRChatMicMuteAutomationsConfig extends AutomationConfig {
+  mode: VRChatVoiceMode;
+  onSleepModeEnable: 'MUTE' | 'UNMUTE' | 'NONE';
+  onSleepModeDisable: 'MUTE' | 'UNMUTE' | 'NONE';
+  onSleepPreparation: 'MUTE' | 'UNMUTE' | 'NONE';
+}
+
 // STATUS AUTOMATIONS
 export interface ChangeStatusBasedOnPlayerCountAutomationConfig extends AutomationConfig {
   limit: number;
@@ -233,6 +244,7 @@ export interface AutoAcceptInviteRequestsAutomationConfig extends AutomationConf
 
 // SHUTDOWN AUTOMATIONS
 export type PowerDownWindowsMode = 'SHUTDOWN' | 'REBOOT' | 'SLEEP' | 'HIBERNATE' | 'LOGOUT';
+
 export interface ShutdownAutomationsConfig extends AutomationConfig {
   triggerOnSleep: boolean;
   sleepDuration: number;
@@ -398,6 +410,13 @@ export const AUTOMATION_CONFIGS_DEFAULT: AutomationConfigs = {
     releaseFootLockOnPoseChange: true,
     footLockReleaseWindow: 600,
     oscScripts: {},
+  },
+  VRCHAT_MIC_MUTE_AUTOMATIONS: {
+    enabled: true,
+    mode: 'TOGGLE',
+    onSleepModeEnable: 'NONE',
+    onSleepModeDisable: 'NONE',
+    onSleepPreparation: 'NONE',
   },
   // STATUS AUTOMATIONS
   CHANGE_STATUS_BASED_ON_PLAYER_COUNT: {