+++ src-ui/app/views/dashboard-view/views/vrchat-mic-mute-automations-view/vrchat-mic-mute-automations-view.component.ts
@@ -1,0 +1,96 @@
+import { Component, DestroyRef, OnInit } from '@angular/core';
+import { SelectBoxItem } from '../../../../components/select-box/select-box.component';
+import { AutomationConfigService } from '../../../../services/automation-config.service';
+import { VRChatMicMuteAutomationsConfig, VRChatVoiceMode } from '../../../../models/automations';
+import { map } from 'rxjs';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+
+@Component({
+  selector: 'app-vrchat-mic-mute-automations-view',
+  templateUrl: './vrchat-mic-mute-automations-view.component.html',
+  styleUrls: ['./vrchat-mic-mute-automations-view.component.scss'],
+})
+export class VRChatMicMuteAutomationsViewComponent implements OnInit {
+  modeOptions: SelectBoxItem[] = [
+    {
+      id: 'TOGGLE',
+      label: 'vrchatMicMuteAutomations.modeOptions.TOGGLE',
+    },
+    {
+      id: 'PUSH_TO_MUTE',
+      label: 'vrchatMicMuteAutomations.modeOptions.PUSH_TO_MUTE',
+    },
+  ];
+  modeOption: SelectBoxItem | undefined;
+  muteActionOptions: SelectBoxItem[] = [
+    {
+      id: 'NONE',
+      label: 'vrchatMicMuteAutomations.muteOptions.NONE',
+    },
+    {
+      id: 'MUTE',
+      label: 'vrchatMicMuteAutomations.muteOptions.MUTE',
+    },
+    {
+      id: 'UNMUTE',
+      label: 'vrchatMicMuteAutomations.muteOptions.UNMUTE',
+    },
+  ];
+  onSleepEnableMuteOption: SelectBoxItem | undefined;
+  onSleepDisableMuteOption: SelectBoxItem | undefined;
+  onSleepPreparationMuteOption: SelectBoxItem | undefined;
+
+  constructor(
+    private automationConfigService: AutomationConfigService,
+    private destroyRef: DestroyRef
+  ) {}
+
+  async ngOnInit() {
+    this.automationConfigService.configs
+      .pipe(
+        takeUntilDestroyed(this.destroyRef),
+        map((configs) => configs.VRCHAT_MIC_MUTE_AUTOMATIONS)
+      )
+      .subscribe((config) => {
+        this.modeOption = this.modeOptions.find((o) => o.id === config.mode);
+        this.onSleepEnableMuteOption = this.muteActionOptions.find(
+          (o) => o.id === config.onSleepModeEnable
+        );
+        this.onSleepDisableMuteOption = this.muteActionOptions.find(
+          (o) => o.id === config.onSleepModeDisable
+        );
+        this.onSleepPreparationMuteOption = this.muteActionOptions.find(
+          (o) => o.id === config.onSleepPreparation
+        );
+      });
+  }
+
+  onChangeModeOption($event: SelectBoxItem | undefined) {
+    if (!$event) return;
+    this.automationConfigService.updateAutomationConfig<VRChatMicMuteAutomationsConfig>(
+      'VRCHAT_MIC_MUTE_AUTOMATIONS',
+      {
+        mode: $event.id as VRChatVoiceMode,
+      }
+    );
+  }
+
+  async onChangeMuteOption(
+    automation: 'ON_SLEEP_ENABLE' | 'ON_SLEEP_DISABLE' | 'ON_SLEEP_PREPARATION',
+    option: SelectBoxItem | undefined
+  ) {
+    if (!option) return;
+    const keyMap = {
+      ON_SLEEP_ENABLE: 'onSleepModeEnable',
+      ON_SLEEP_DISABLE: 'onSleepModeDisable',
+      ON_SLEEP_PREPARATION: 'onSleepPreparation',
+    };
+    const key = keyMap[automation];
+    await this.automationConfigService.updateAutomationConfig<VRChatMicMuteAutomationsConfig>(
+      'VRCHAT_MIC_MUTE_AUTOMATIONS',
+      {
+        [key]: option!.id as 'MUTE' | 'UNMUTE' | 'NONE',
+      }
+    );
+  }
+}